move get_file to helpers module
