+++ ezproxylookup/__init__.py
@@ -3,3 +3,4 @@ from flask import Flask
 app = Flask(__name__)
 
 import ezproxylookup.views
+import ezproxylookup.helpers
